remove the existing rt module and refactor visitor
